+++ pages/portfolio/index.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from "next";
 
 // Import components
-import PortofolioPageV2 from "../../components/v1/PortofolioPage/PortofolioPageV2";
+import PortfolioPage from "../../components/v1/PortfolioPage/PortfolioPageV2";
 
 const Home: NextPage = () => {
-    return <PortofolioPageV2 />;
+    return <PortfolioPage />;
 };
 
 export default Home;
